+++ SW/visualisation/src/src/main/java/Utils/Vector2D.java
@@ -4,27 +4,27 @@ package Utils;
  * Created by secured on 2016. 03. 06..
  */
 public class Vector2D {
-    private int _coordinateX;
-    private int _coordinateY;
+    private float _coordinateX;
+    private float _coordinateY;
 
-    public Vector2D(int coordinateX, int coordinateY) {
+    public Vector2D(float coordinateX, float coordinateY) {
         _coordinateX = coordinateX;
         _coordinateY = coordinateY;
     }
 
-    public int get_coordinateX() {
+    public float get_coordinateX() {
         return _coordinateX;
     }
 
-    public void set_coordinateX(int coordinateX) {
+    public void set_coordinateX(float coordinateX) {
         _coordinateX = coordinateX;
     }
 
-    public int get_coordinateY() {
+    public float get_coordinateY() {
         return _coordinateY;
     }
 
-    public void set_coordinateY(int coordinateY) {
+    public void set_coordinateY(float coordinateY) {
         _coordinateY = coordinateY;
     }
 }
